For FB, prefer internal build over global build
Summary:
This commit makes it so when running the internal version of Flipper, we always prefer the internal build of React DevTools as opposed to a global install of DevTools. Note that users may still change the version the use and select the global version.

Externally, i.e. for the OSS build, users will still have the option to always prefer the global build and toggle between it.

The motivation for this is to let us ensure that employees are using the latest version of React DevTools and in order to obtain accurate logs of internal usage of DevTools

Reviewed By: mweststrate

Differential Revision: D33167394

fbshipit-source-id: ee75eedc3a488e223189bbfa6288408bb84824e3