Fix docs site resource load errors
Summary: I noticed an error on loading "google-anaytics" script, "code-blocks-buttons" script and "network.png" screenshot (which is supposed to be a part of carousel on the main page) in dev console when Flipper docs website is opened. This diff fixes them. Looks like it positively affects load speed as well because each of these failed requests takes 1-2 seconds to resolve before the failure status returned.

Reviewed By: mweststrate

Differential Revision: D33313293

fbshipit-source-id: a1f6357d89121eefa2df9aeae783cd48bc3beaf5