Adresses an issue with menu mutation
Summary:
A Logview issue was raised for the following error:
Cannot assign to read only property '0' of object '[object Array]'.

Analysis:
Currently, the menuEntries are stored by reference in the reducer, and made read only to protect it against accidental writes. Tha's probably why the '0' assignment fails, and the fix is that the reducer should only get a defensive copy as config about the current menu entries to show.

Reviewed By: mweststrate

Differential Revision: D33454514

fbshipit-source-id: 2faf3036eb6ae4fbe9a8d0253fde27cf389d3e4c