Fix windows build (for real this time?) (#3289)
Summary:
This previous attempt didn't work. Not even sure what happens now, it seems to just return the result of a boolean evaluation? Cool.

Still haven't verified that this works in the release build but noticed that we do effectively the same in the build step. By upgrading to the same revision of the action we use and then switching from `&&` to simply `;`, we can be reasonably sure that this will work now.

The alternative would be to switch to `pwsh` which appears to be the new name. `powershell` is the default shell and seems to be an older version without support for `&&`. I honestly never want to touch this or read about it again, so let's go with the simpler option.

Pull Request resolved: https://github.com/facebook/flipper/pull/3289

Reviewed By: aigoncharov

Differential Revision: D33582853

Pulled By: passy

fbshipit-source-id: 7c70ff360a5c327139b6ecb5463e85883a2caf9d