Move idb iOSBridge free-functions into class
Summary:
This isn't that important right now, but will help as the `IOSBridge` is expanded in the future. There's also an argument for logical ordering of functions within a container (a class!), but that's not the real motivation here.

The existence of `bind` shows that this free function is effectively closing over this argument anyway.

Reviewed By: passy, lblasa

Differential Revision: D33818267

fbshipit-source-id: e7b83f013121cedbd31cb28746b69c1fa76a0803