Hopefully fix flipper-release npm publish step (#3366)
Summary:
This is really stupid but the download creates a directory with the same name before putting the file in it. I think `v2` no longer does but hey, this is the devil we know.

You can look at the actions further down to see that we reference all the other downloads by `filename/filename` as well, so this should do it.

Pull Request resolved: https://github.com/facebook/flipper/pull/3366

Test Plan: Another release, I'm afraid. This is a step you cannot dry-run.

Reviewed By: cekkaewnumchai, aigoncharov

Differential Revision: D33842700

Pulled By: passy

fbshipit-source-id: d6f7a4a711ec056ba16f13c72f2c550e9a632ac0