Use idb to record devices
Summary:
`idb` has supported video recording on devices for quite some time now. We even have the implementation in `iOSDevice`, but it's currently unused.

As a result, we just need to remove the check within `iOSDevice` that the target is an "emulator" and it will use the `idb` implementation. Removing this check means that the button is now selectable within the UI.

Reviewed By: passy

Differential Revision: D33913383

fbshipit-source-id: b3d6b4c480c79b54f0d2c206c68092ab62bf71ed