Fix build failure on OSS master (#3394)
Summary:
Pull Request resolved: https://github.com/facebook/flipper/pull/3394

This diff caused OSS builds failing due to a FB plugin that requires emotion, instead of using it as peer dep. This introduced a duplicate dep, which doesn't get installed in OSS, causing patch-package for that package to fail. Changed it to a peer dep and removed the patch.

Reviewed By: aigoncharov

Differential Revision: D33975202

fbshipit-source-id: 1e82964988e944a28ecb46a10127ad9e13f743f7