Release client if exists before establishing a connection
Summary:
The problem seems to be exclusive to Android emulators.

Bringing the emulator ON from a warm state brings the app from is previous state before shutting down. At this point there may have been a Flipper connection.

Flipper tries to connect and does so successfully replacing the previously, now dead connection.

The problem is that replacing that connection triggers the reconnect cycle as Flipper thinks the connection is dead and hence the reconnect loop initiates.
Changelog: Release an existing client before attempting a secure connection

Reviewed By: passy

Differential Revision: D34080726

fbshipit-source-id: 8185adb492dd4d9255fcea5874ca2e5b7fee0c84