Specialize error message when no xcode selected
Summary:
We can detect an xcode-select failure earlier on, right now this will be misleading as later healtchecks around `xctrace`/`xcodebuild` will fail and it's not clear why.

If you don't have any Xcode selected, then we can detect this and mark it as an error. This can be done by looking out the output of the command.

To do this we have to propogate the stdout of the shell-out back, so there's a subtype of the regular healtcheck. This is then inspected for it's path since `xcode-select` returns the same exit code if xcode isn't selected or at it's default state.

Reviewed By: passy

Differential Revision: D34168673

fbshipit-source-id: c0aa846bb251aaf026beb976a042b727c7cf1c24