Add library verification step (#3436)
Summary:
This is pretty dumb but hopefully good enough to prevent accidental regressions. We simply check if there are "enough" `libevent_core.so`s in the bundle. This is obviously not future-proof but it's super cheap to run and if it causes problems at some point, we can always remove it.

Apologies for the formatting spam. Didn't notice until I submitted this that my editor got a little passive-aggressive.

Pull Request resolved: https://github.com/facebook/flipper/pull/3436

Test Plan: CI here.

Reviewed By: lblasa

Differential Revision: D34210743

Pulled By: passy

fbshipit-source-id: a57c397e39456fae33af9f3ceed08b6944eac79e