Refactor recommended plugin API
Summary:
This is just refactoring in preparation to open-sourcing internal plugin ditribution code to make it avaialble for other orgs.

This diff moves recommended plugins handling from `pluginMarketplace` which will be the same for fb-internal and OS versions into `pluginMarketplaceAPI` which will differ for fb-internal and OS versions. This will make it possible for other orgs to define their own "recommended" plugins which then will be automatically installed/enabled for new users.

Differential Revision: D34379981

fbshipit-source-id: 630aedac1cf8030b5921a6f3628aefdab1cdd0eb