Update babel config
Summary:
allow-large-files

During decapitation the style guide was accidentally removed. In this diff we bring it back and embed it into our docs.
The only caveat is that the fonts have slightly changes because we included antd.css. It does not change too much from my point of view, but I would love to hear from the team.

Reviewed By: jknoxville

Differential Revision: D34339758

fbshipit-source-id: 10d347bc805f9314ae717de483bf8b766000280f