docs: add social banner in support of Ukraine (#3496)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156651121-b1479b59-a4e7-4a5c-b8d5-4265aa0aea7a.png)

Pull Request resolved: https://github.com/facebook/flipper/pull/3496

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: staticdocs**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34622802/V3/staticdocs/)|

|**Modified Pages**|

**Static Docs Preview: flipper**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34622802/V3/flipper/)|

|**Modified Pages**|

Reviewed By: mweststrate

Differential Revision: D34622802

Pulled By: passy

fbshipit-source-id: a18c0de24a6a6c2311aef9a046fe65d917ce6f9d