Keytar platform/arch
Summary:
There appears to be an issue with login on M1 Macs right now. To address this, this resolves the native binary based on a `(platform, arch)` tuple.

The binaries are from here: https://github.com/atom/node-keytar/releases/tag/v7.9.0

The binary size increase here is notable so as a follow-up we could remove the non-relevant binaries as part of the packing process.

Reviewed By: aigoncharov

Differential Revision: D34685236

fbshipit-source-id: d860a59517a769a59055dfca24d01baecd8d1430