Fix expired tokens showing as being correctly logged in
Summary: In the category "how did this ever work", we had quite some reports of Flipper features not working while users are seemingly logged in. The cause of that is because the user info gets stored in the user reducer, so even when the network requests fail, the users appear to be logged in.

Reviewed By: nikoant

Differential Revision: D35114048

fbshipit-source-id: b8c09cb4e3fd7cd0cbe87ecb06dafca88716b689