Fix flipper-server build
Summary:
flipper-server build fails when a server add-on uses one of the ignored modules.
For instance, React DevTools is goin got use `ws` which has an optional dependency on `utf-8-validate` and `bufferutil`. Even though it is optional, it is still required, and therefore picked up by Metro.
This way it is silently ignored.

Reviewed By: mweststrate

Differential Revision: D34716514

fbshipit-source-id: a9d463d2d5d7c954e35d9a16f3da44ce17615673