Script for platform-specific flipper-server bundles
Summary:
To roll out Flipper Server for desktop, we need a bundle that comes with all node dependencies pre-installed and a bundled Node runtime. This creates some platform-specific sub-folders in `dist/` with both.

The `--mac`, `--linux`, `--win` options are chosen to be compatible with the main build script. For now, we only build x64 builds for Mac which is also in line with the Desktop build as we don't have signing for the whole bundle.

Reviewed By: lblasa

Differential Revision: D35545492

fbshipit-source-id: cce7165916d91a333f305713b9d6d7b9984984f4