Fix native tasks prepare ordering
Summary:
^

This change makes it clearer the order of tasks for the native libraries.

If unspecified, tasks will run in parallel.

There was an additional issue in which the revision could be written to disk before all dependencies are properly downloaded and packaged.

For reference, we depend on Folly.
Folly depends on:
- Double-Conversion
- Glog
- Boost
- LibEvent

This is now reflected in the tasks graph.

Reviewed By: passy

Differential Revision: D35931187

fbshipit-source-id: ec19a28521ebf318bd1e92feafab1671733679ca