Extract flipper server instantiation
Summary:
^

This is a very small refactoring and addition to the Flipper Desktop app.

The instantiation of FlipperServer was extracted to a separate function.

This a very tiny change that allows to switch the implementation we use for FlipperServer in a more convenient way.

In this same change, an unused function is added which will create a separate FlipperServer instance that uses a web-socket underneath which makes it possible to use with flipper-server.

So, if interested, it is enough to call that function instead of the one currently in use and it will make Flipper connect to flipper-server instead.

https://pxl.cl/24j8R

Reviewed By: passy

Differential Revision: D36440574

fbshipit-source-id: 94ea2ab7208b898a82ac5e7fd7edd9cb824b4810