Fix height of launch emu window
Summary:
A paper cut that hurt me so many times. The launch emulator window gets populated by external commands that finish at interminate points in time and change the size, jiggling the buttons around. That means, you may hover over an android emulator just when the iOS command returns and when you click, you don't hit the Android emu, but some iOS thing.

This just makes the height fixed, so this cannot happen.

Still having some more improvements in mind as this is such a high-trafficked surface.

Chanlog: Fix wiggling of the Launch Emulator dialogue as data comes in

Reviewed By: mweststrate

Differential Revision: D36595001

fbshipit-source-id: e61de55f348ab4d2f6f977552539d2a44d3fd11d