Change server template
Summary:
This is a completely meaningless change *but* it actually has an important
semantic meaning. I recently changed the `+x` bit to be set on the node binary
while bundling which is not always true when packing on Linux.

However, we don't consider permission bits as part of the checksums we build,
which means you will always get the previous version until we update the node version
or this very script because it's part of the same bundle.

So let's just ship this "update".

Reviewed By: mweststrate

Differential Revision: D36954703

fbshipit-source-id: 2d89cb1ee2e7bcce125c3f9efc9d8d39101ac133